CVSL-132: Rename Prisoner search mock API
